requirements.txt
streamlit
google-generativeai
deep-translator
opencage
requests
Pillow
requests
python-dotenv
flask
flask-wtf
streamlit
google-generativeai
deep-translator
opencage
requests
Pillow
requests
python-dotenv
flask
flask-wtf
flask_cors
langchain-openai
langchain_community
langchain
openai
faiss-cpu
pypdf
cryptography<48.0.0


